refactor(core): move checkDataSet to SchemaReader base for CSV/TsFile reuse
